+++ opendc-simulator/opendc-simulator-network/src/main/kotlin/org/opendc/simulator/network/components/node/Node.kt
@@ -56,7 +56,7 @@ internal interface Node<Self : Node<Self>> : IInvalidatable, Msgable<Node<*>>, L
      * ID of the node. Uniquely identifies the node in the [Network].
      */
     val id: NodeId
-        get() = NodeId(ip.intValue().toLong())
+        get() = NodeId(ip.longValue().toUInt())
 
     /**
      * Port speed in Kbps full duplex.
